Rebase Customer to User
Rebased the customer object to a user object for more generality. This way I can also create a string of roles attached to each user that will define their high-level abilities (i.e. CUSTOMER, SERVICER, ADMIN, etc.), but this is yet to be implemented. This rebase only accomplished the task of setting up to combine the two roles into one.