Massive User Authentication Cleanup/Improvement/New Features
Added the ability for users to reset their password, resend email
verifications, logout and delete their account. Registration emails will
be attached with a url that querys the graphql resolvers to setup
verficiation. Nodemailer is the current package of choice for sending said
emails, although may switch in the future.

Issues:
    - I believe I overwrote the package.json while experimenting with
      foundation for email templates. In the next commit I will check
      in all the proper missing dependencies.
    - Unsure how I am going to handle the client after their address has
      been verified.
    - Currently leaking sensitive database and server information
      through thrown Errors.
    - Need to check for potential vulnerabilities in verification system
      that could be exploited
    - Graphql server would not run without contentSecurityPolicy
      disabled, which opens the application up to potential