Solving GraphQL N+1
I added a single userLoader which allows me to cache the results of the
user and preform a single database query on an array of objects. I have
also begun preparing the appliction for testing with the jest and
apollo-server-testing packages. This meant moving the server start out
of the server.js file and into a new file named index.js in the same
directory. I also will be used a unique database for testing purposes.
Lastly, I made a small server update to include cors references.