Finally Added Testing!
I've finally began adding testing to the repo. I am using jest and
apollo-server-testing to do all testing of my graphql resolvers.

In order to begin testing, I had to further separate the server
configuration from the actual startup of the server. This allowed me
more control when testing in stopping the server. I may turn the stop
and start of the server into a global startup independent of the testing
files later on though.

I also have modified many of the resolvers to return data relevant only
when is necessary so as not to clog the output when testing. I am likely
going to move around some of the code involving querying of the server
directly when giving a token to the frontend because I do not
necessarily want direct access to the server at the moment.